Changed result of vg->getExtentSize() from uint64_t into long long to match everything else. vgchange and vgmerge needed some variables changes to match to prevent signed from being compared to unsigned.

+++ trunk/kvpm/volgroup.cpp
@@ -276,7 +276,7 @@ void VolGroup::processPhysicalVolumes(vg_t lvmVG)
 
         for (int x = 0; x < m_member_pvs.size(); x++) {
             if (m_member_pvs[x]->isAllocatable())
-                m_allocatable_extents += m_member_pvs[x]->getRemaining() / (long long) m_extent_size;
+                m_allocatable_extents += m_member_pvs[x]->getRemaining() / m_extent_size;
             m_mda_count += m_member_pvs[x]->getMdaCount();
         }
